Renamed Config Beans
