+++ framework/orchestrator/src/main/java/org/intueri/orchestrator/OrchestratorConfig.java
@@ -16,7 +16,7 @@ import java.util.UUID;
 @Configuration
 @PropertySource("classpath:application.yml")
 @ConfigurationProperties(prefix = "intueri")
-public class ApplicationConfig {
+public class OrchestratorConfig {
 
     private String bootstrapServer = "kafka:9092";
