+++ framework/output/src/main/java/org/intueri/output/OutputConfig.java
@@ -1,0 +1,4 @@
+package org.intueri.output;
+
+public class OutputConfig {
+}
