+++ framework/util/src/main/java/org/intueri/util/IntueriUtil.java
@@ -1,28 +1,12 @@
 /**
  * Copyright (C) 2019 ${author} <speckij@posteo.net>
- *
+ * <p>
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
  */
 package org.intueri.util;
 
-import org.apache.kafka.common.serialization.Serdes;
-import org.apache.kafka.streams.StreamsConfig;
-
-import java.util.Properties;
-
 @SuppressWarnings("CheckStyle")
 public class IntueriUtil {
 
-    public static Properties kafkaProperties(String id, String server) {
-        Properties properties = new Properties();
-        properties.put(StreamsConfig.APPLICATION_ID_CONFIG, id);
-        properties.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, server);
-        properties.put(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, Serdes.String().getClass());
-        properties.put(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, Serdes.String().getClass());
-        final String serializer = "org.apache.kafka.common.serialization.StringSerializer";
-        properties.put("key.serializer", serializer);
-        properties.put("value.serializer", serializer);
-        return properties;
-    }
 }
